Move function to helper file
